+++ cpp/src/binaryop/jit/code/operation.cpp
@@ -450,6 +450,22 @@ const char* operation =
             return PMod::operate<TypeOut, TypeRhs, TypeLhs>(y, x);
         }
     };
+
+    struct ATan2 {
+        template <typename TypeOut, typename TypeLhs, typename TypeRhs>
+        static TypeOut operate(TypeLhs x, TypeRhs y) {
+            return TypeOut{std::atan2(double{x}, double{y})};
+        }
+    };
+
+    struct RATan2 {
+        template <typename TypeOut,
+                  typename TypeLhs,
+                  typename TypeRhs>
+        static TypeOut operate(TypeLhs x, TypeRhs y) {
+            return TypeOut{ATan2::operate<TypeOut, TypeRhs, TypeLhs>(y, x)};
+        }
+    };
 )***";
 
 }  // namespace code