+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - PR #4992 Add Java bindings for converting nans to nulls
 - PR #4975 Add Java bindings for first and last aggregate expressions based on nth
 - PR #5036 Add positive remainder binary op functionality
+- PR #5055 Add atan2 binary op
 - PR #5099 Add git commit hook for clang-format
 - PR #5072 Adding cython binding to `get_element`
 
@@ -147,6 +148,7 @@
 - PR #5093 Add `.cat.as_known` related test in `dask_cudf`
 - PR #5104 Add missing `.inl` files to clang-format and git commit hook
 - PR #5101 Add POSITION_INDEPENDENT_CODE flag to static cudftestutil library
+- PR #5109 Update CONTRIBUTING.md for `clang-format` pre-commit hook
 
 ## Bug Fixes